minor improvement on HTTP factory

+++ app/scripts/http/API.js
@@ -7,6 +7,9 @@ angular.module('angular.bootstrap')
 
 	return {
 		get:function(url){
+			if(!url){
+				throw "url has to be defined";
+			}
 			return $http.get(baseUrl+url);
 		}, 		
 		post:function(url, bodyParams){
@@ -17,10 +20,16 @@ angular.module('angular.bootstrap')
 		}, 
 
 		put:function(url, bodyParams){
+			if(!bodyParams || !angular.isObject(bodyParams)){
+				throw "Second parameters need to be an Object";
+			}
 			return $http.put(baseUrl+url,bodyParams);
 		}, 
 
 		delete: function(url){
+			if(!url){
+				throw "url has to be defined";
+			}
 			return $http.delete(url);
 		}	
 	};
